version 0.0.22

+++ src/pylejandria/upload.py
@@ -1,6 +1,6 @@
 """
 This is a simple tool to upload the Pylejandria package to Pypi and GitHub with
-a UI. Version 1.0.3 By Armando Chaparro.
+a UI. Version 1.0.5 By Armando Chaparro.
 """
 
 import os
@@ -18,10 +18,16 @@ class Uploader:
         the package to Pypi and GitHub, is easier than parse all the terminal
         arguments.
         """
+        self.root = tk.Tk()
+        self.root.title('Uploader By Armando Chaparro')
+        self.root.wm_protocol("WM_DELETE_WINDOW", self.quit)
+        self.path_entry = tk.Entry(self.root, width=50)
+        self.version_entry = tk.Entry(self.root, width=15)
+        self.commit_entry = tk.Entry(self.root, width=30)
+        self.git_combobox = ttk.Combobox(self.root, width=15)
         self.regex = '[0-9]+\.[0-9]+\.[0-9]+'
-        self.path = os.getcwd()
-        self.get_version()
-        self.commit = f'version {self.version}'
+        if not self.change_path():
+            self.quit()
 
     def get_version(self) -> None:
         """
@@ -30,10 +36,18 @@ class Uploader:
         """
         with open(os.path.join(self.path, 'setup.cfg'), 'r') as f:
             self.text = f.read()
-            self.old_version = re.search(self.regex, self.text).group()
+            if match := re.search(self.regex, self.text):
+                self.old_version = match.group()
+            else:
+                return
             match self.old_version.split('.'):
                 case(x, y, z):
                     self.version = f'{x}.{y}.{int(z) + 1}'
+        self.commit = f'version {self.version}'
+        self.commit_entry.delete(0, tk.END)
+        self.commit_entry.insert(0, self.commit)
+        self.version_entry.delete(0, tk.END)
+        self.version_entry.insert(0, self.version)
 
     def validate_version(self) -> None:
         """
@@ -42,6 +56,7 @@ class Uploader:
         """
         text = self.version_entry.get()
         if re.match(f'^{self.regex}$', text):
+            self.version = text
             self.version_status['fg'] = '#00ff00'
             self.version_status['text'] = 'Valid version'
         else:
@@ -92,8 +107,20 @@ class Uploader:
         """
         tk.Tk().withdraw()
         self.path = askdirectory()
+        if not self.path:
+            return False
         self.path_entry.delete(0, tk.END)
         self.path_entry.insert(0, self.path)
+        git_repo = os.path.exists(os.path.join(self.path, '.git'))
+        self.commit_entry['state'] = 'normal' if git_repo else 'disabled'
+        self.git_combobox['state'] = 'readonly' if git_repo else 'disabled'
+        try:
+            self.get_version()
+        except FileNotFoundError:
+            self.version_entry.delete(0, tk.END)
+            self.commit_entry.delete(0, tk.END)
+        return True
+
 
     def update_git(self, event: tk.Event) -> None:
         """
@@ -121,57 +148,51 @@ class Uploader:
         """
         Main function of Uploader, it creates all the UI and bindings.
         """
-        root = tk.Tk()
-        root.title('Uploader By Armando Chaparro')
-
-        path_label = tk.Label(root, text='Folder')
+        path_label = tk.Label(self.root, text='Folder')
         path_label.grid(row=0, column=0, padx=5, sticky='w')
-        self.path_entry = tk.Entry(root, width=50)
-        self.path_entry.insert(0, self.path)
         self.path_entry.grid(row=0, column=1, padx=5, sticky='w')
-        path_button = tk.Button(root, text='Change', command=self.change_path)
+        path_button = tk.Button(self.root, text='Change', command=self.change_path)
         path_button.grid(row=0, column=2, padx=5, sticky='w')
 
-        version_label = tk.Label(root, text='Version')
+        version_label = tk.Label(self.root, text='Version')
         version_label.grid(row=1, column=0, padx=5, sticky='w')
-        self.version_entry = tk.Entry(root, width=15)
-        self.version_entry.insert(0, self.version)
         self.version_entry.grid(row=1, column=1, padx=5, sticky='w')
         version_button = tk.Button(
-            root, text='Validate', command=self.validate_version
+            self.root, text='Validate', command=self.validate_version
         )
         version_button.grid(row=1, column=2, padx=5)
-        self.version_status = tk.Label(root, text='')
+        self.version_status = tk.Label(self.root, text='')
         self.version_status.grid(row=1, column=3, padx=5)
 
-        git_label = tk.Label(root, text='Upload to GIT')
+        git_label = tk.Label(self.root, text='Upload to GIT')
         git_label.grid(row=2, column=0, padx=5, sticky='w')
-        self.git_combobox = ttk.Combobox(root, width=15)
         self.git_combobox['values'] = ['True', 'False']
-        self.git_combobox['state'] = 'readonly'
         self.git_combobox.bind("<<ComboboxSelected>>", self.update_git)
         self.git_combobox.current(0)
         self.git_combobox.grid(row=2, column=1, padx=5, sticky='w')
 
-        commit_label = tk.Label(root, text='Commit')
+        commit_label = tk.Label(self.root, text='Commit')
         commit_label.grid(row=3, column=0, padx=5, sticky='w')
-        self.commit_entry = tk.Entry(root, width=30)
-        self.commit_entry.insert(0, self.commit)
         self.commit_entry.grid(row=3, column=1, padx=5, sticky='w')
 
-        pypi_label = tk.Label(root, text='Upload to Pypi')
+        pypi_label = tk.Label(self.root, text='Upload to Pypi')
         pypi_label.grid(row=4, column=0, padx=5, sticky='w')
-        self.pypi_combobox = ttk.Combobox(root, width=15)
+        self.pypi_combobox = ttk.Combobox(self.root, width=15)
         self.pypi_combobox.bind("<<ComboboxSelected>>", self.update_pypi)
         self.pypi_combobox['values'] = ['True', 'False']
         self.pypi_combobox['state'] = 'readonly'
         self.pypi_combobox.current(0)
         self.pypi_combobox.grid(row=4, column=1, padx=5, sticky='w')
 
-        upload_button = tk.Button(root, text='Upload', command=self.get_values)
+        upload_button = tk.Button(self.root, text='Upload', command=self.get_values)
         upload_button.grid(row=5, column=1, padx=5, pady=10, sticky='ew')
 
-        root.mainloop()
+        self.root.mainloop()
+    
+    def quit(self) -> None:
+        self.root.destroy()
+        exit()
+
 
 if __name__ == '__main__':
     app = Uploader()
